Updated parsing algorithm.
Added a new function that will analyze the vector of parsed tokens. This
is to account for strings that have connectors inside of them. I'm not
sure if the parse function is necessary anymore, or if we can merge both
functions for a more efficient algorithm.

Some tests were made, but I still need to check other cases. Also found
a potential problem with having multiple connectors put together. Will
address issue later.

Comments were made for both functions.